docker/lab-data-provider/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
httpx>=0.25.0
tavily-python>=0.5.0
